Allow entries to be force pulled from WP (GH #8)
If the local page cache for a blog is invalid, the component
will continue to load data from the local page cache, meaning
the blog won't appear to update in Joomla. This update allows
a URL parameter to be added to an entries view on the Joomla
frontend that forces a feed request to WP to get the page.
To use, add 'nocache=true' to the entries view URL.