Be smarter about closing content (GH #22)
When a read more link is put in, HTML elements earlier in the
content must be closed, so it's not suitable to simply put in
a closing </p> and hope it works. This change strips out <p></p>
and <div></div> content after the break - hopefully leaving
closing elements in place.