Allow WP image links to be replaced (GH #25)
WordPress.com feeds may include images linking to things like
digg, adding a comment, twitter, etc. These appear as blocky
text boxes, which are not very nice. This update adds some
optional javascript which converts theses images to text, and
adds the CSS class 'wordbridge_wp_link' to the link a element.
This will allow the links to be styled.

To change links, set the 'Convert Wordpress links to text'
parameter on the blog menu item to 'Yes'